FEM: results, move principal and shear calculation from result reader to result tools and use new result stress float lists
