+++ src/Mod/Fem/feminout/importCcxFrdResults.py
@@ -105,6 +105,7 @@ def importFrd(filename, analysis=None, result_name_prefix=None):
                     res_obj = restools.compact_result(res_obj)
                 res_obj = restools.add_disp_apps(res_obj)  # fill DisplacementLengths
                 res_obj = restools.add_von_mises(res_obj)  # fill StressValues
+                res_obj = restools.add_principal_stress(res_obj)  # fill PrincipalMax, PrincipalMed, PrincipalMin, MaxShear
                 res_obj = restools.fill_femresult_stats(res_obj)  # fill Stats
         else:
             error_message = (
